requirements.txt
numpy==2.2.3
pillow==11.1.0
sentencepiece==0.2.0
torch==2.6.0
absl-py==2.1.0